+++ carball/analysis/events/hit_detection/hitbox/car.py
